+++ Program.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using AdressBook.files; //this is the namespace for the files that are in the project. This is where the classes are stored
 
-namespace AdressBook
+namespace snackShack
 {
     internal static class Program
     {
@@ -18,11 +17,10 @@ namespace AdressBook
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new frm_addressForm());
+            Application.Run(new frm_main());
         }
         internal const bool debug = true; //global var to controll default for debugs
-        internal static List<Contact> contacts = new List<Contact>(); //makes a collection that refferences the class that will store information
-        internal const string preMadeErrorMsg = "I'm sorry dave, I'm afraid I can't do that";
+        internal static List<inventory> inventory = new List<inventory>(); //makes a collection that refferences the class that will store information
     }
     internal static class coreCommands
     {
@@ -37,5 +35,20 @@ namespace AdressBook
         } //this is for error messages. the text is mandatory, but the
           //rest isn't necessary. if you don't provide an exception, it won't display one. If you provide an error, it will default to "hide error".
           // WARNING: If you provide a boolean, you can controll if it's hidden. If you provide a boolean but no error, well, it won't work righ
-    }
+
+        internal static string path()
+        {
+            var pathSpliter = '\\'; //part between paths
+            var loc = AppDomain.CurrentDomain.BaseDirectory; //get the current domain
+            var comps = loc.Split('\\'); //array of path after being split
+            var len = comps.Length; //length of path
+            var completeLoc = string.Empty; //blank variable for path
+            for (var i = 0; i < len-3; i++) //repeat by the length minus 2, which will lead to it excluding the 2 folders that are unwanted
+            {
+                var comp = comps[i];
+                completeLoc = completeLoc + comp + char.ToString(pathSpliter); //add next part to path
+            }
+            return completeLoc; //output path
+        } //derive the path of the base file
+    } //base commands to aid in everything
 }
